EnemiesGeneration 1.01V
Now each spawn point creates its own type of enemies and has a goal, that is, an object that the enemies are heading towards. At the moment of creation, the spawn point transmits its target to the enemy so that he heads towards it. There should be several such goals, each spawner has its own goal.

The goals themselves move along a pre-determined route, from one point to another.